minor fixes

+++ pkg/models/addBook.go
@@ -11,7 +11,7 @@ func AddBook(title string, author string, quantity int) (string, error){
 		return "", err
 	}
 
-	bookExists, book, err := bookExists(title)
+	bookExists, book, err := bookExists(title, author)
 	if err != nil {
 		return "", err
 	}
@@ -36,7 +36,7 @@ func AddBook(title string, author string, quantity int) (string, error){
 	}
 }
 
-func bookExists(title string) (bool, types.Book, error) {
+func bookExists(title string, author string) (bool, types.Book, error) {
 	var book types.Book
 
 	db, err := Connection()
@@ -44,8 +44,8 @@ func bookExists(title string) (bool, types.Book, error) {
 		return false, book, err
 	}
 
-	selectquery := `SELECT * FROM books WHERE title = ?;`
-	err = db.QueryRow(selectquery, title).Scan(&book.BookId, &book.Title, &book.Author, &book.Available, &book.Quantity)
+	selectquery := `SELECT * FROM books WHERE title = ? AND author = ?;`
+	err = db.QueryRow(selectquery, title, author).Scan(&book.BookId, &book.Title, &book.Author, &book.Available, &book.Quantity)
 	if err != nil {
 		if err != sql.ErrNoRows {
 			return false, book, err
@@ -55,10 +55,62 @@ func bookExists(title string) (bool, types.Book, error) {
 	return true, book, nil
 }
 
+func RemoveBooks(id int, remQuantity int) (string, error) {
+	db, err := Connection()
+	if err != nil {
+		return "", err
+	}
+
+	checkquery := `SELECT available, quantity FROM books WHERE id = ?;`
+	var available, oldQuantity int
+	err = db.QueryRow(checkquery, id).Scan(&available, &oldQuantity)
+	if err != nil {
+		return "", err
+	}
+	if (remQuantity > available) {
+		return "Not enough books available", nil
+	}
+
+	newQuantity := oldQuantity - remQuantity
+	newAvailable := available - remQuantity
+
+	updatequery1 := `UPDATE books SET available = ? WHERE id = ?;`
+	_, err = db.Exec(updatequery1, newAvailable, id)
+	if err != nil {
+		return "", err
+	}
+	updatequery2 := `UPDATE books SET quantity = ? WHERE id = ?;`
+	_, err = db.Exec(updatequery2, newQuantity, id)
+	if err != nil {
+		return "", err
+	}
+	return "Books removed successfully", nil
+}
+
 func DeleteBook(id int) (string, error) {
 	db, err := Connection()
 	if err != nil {
 		return "", err
+	}
+
+	checkPending := `SELECT COUNT(*) FROM checkouts WHERE book_id = ? AND status = 'pending';`
+	var pending int
+	err = db.QueryRow(checkPending, id).Scan(&pending)
+	if err != nil {
+		return "", err
+	}
+	if (pending > 0) {
+		return "Cannot delete book with pending requests", nil
+	}
+
+	checkAvailable := `SELECT available and quantity FROM books WHERE id = ?;`
+	var available, quantity int
+	err = db.QueryRow(checkAvailable, id).Scan(&available, &quantity)
+	if err != nil {
+		return "", err
+	}
+	if (available != quantity) {
+		return "Cannot delete book with pending requests", nil
 	}
 
 	deletequery := `DELETE FROM books WHERE id = ?;`
